radiomanager-frontend: Introduce LibraryPage store hook
