+++ services/radiomanager-frontend/src/components/entries/LibraryPage/hooks/useHandleLibraryLastUploadedTrack.ts
@@ -1,0 +1,16 @@
+import { useEffect } from 'react'
+import { UserTrack } from '@/api/api.types'
+import { useMediaUploader } from '@/modules/MediaUploader'
+
+export const useHandleLibraryLastUploadedTrack = (
+  onLastUploadedTrack: (track: UserTrack) => void,
+) => {
+  const { lastUploadedTrack } = useMediaUploader()
+  useEffect(() => {
+    if (!lastUploadedTrack) {
+      return
+    }
+
+    onLastUploadedTrack(lastUploadedTrack.track)
+  }, [lastUploadedTrack, onLastUploadedTrack])
+}
